fix: выносит mockState в отдельный файл
